pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "assert-review"
version = "0.1.0"
requires-python = ">=3.12"
description = "AI-powered code review interface"
readme = "README.md"

[tool.hatch.build.targets.wheel]
packages = ["ml", "apps/api"]

[project.optional-dependencies]
api = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "httpx>=0.27.0",
]
ml = [
  "torch>=2.2.0",
  "transformers>=4.40.0",
  "datasets>=2.19.0",
  "faiss-cpu>=1.8.0",
  "hdbscan>=0.8.33",
  "scikit-learn>=1.4.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "wandb>=0.17.0",
  "onnx>=1.16.0",
  "onnxruntime>=1.18.0",
  "optimum>=1.19.0",
  "peft>=0.11.0",
  "tqdm>=4.66.0",
  "umap-learn>=0.5.6",
  "scipy>=1.13.0",
]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["apps/api/tests", "ml/data/tests", "ml/models/tests", "ml/eval/tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
